aqualink: begin daemon implementation in C
There are open architectural questions about the control daemon. Will
it be writtem in C? Will it be linux-specific? What IPC library will
it use? What is the monitoring and control IPC protocol?

Instead of debating indefinitely, leave these questions open, and
begin a C implementation. Base the event loop on libubox, as it saves
us from having to call poll() directly. Because libubox is not
provided by most distros, build a local copy by default.

Periodically send the GET_MEASUREMENTS command and decode the replies.
The main purpose is to establish comms, and make sure the RS-485 link
and packet decoder work.

Signed-off-by: Alexandru Gagniuc <mr.nuke.me@gmail.com>